add key_id.py script
This script accepts a list of notes as args, and will output a list of
keys that contain all of these notes. If there is no output, there were
no keys that contained all of these notes. This does not include keys
built upon sharps. It will expect only normal or sharp notes instead of
flats.