Complete rewrite
Previously, keygen was very complicated and unmaintainable. It was
written before I had any development experience. It was simply trying to
do too much. It also relied on an older version of midi_abstraction
which I've since rewritten to have a more reasonable interface (as well
as fixing some bugs that caused it to include incorrect notes).

This commit does the following:
* Delete the old project in entirety
* Rewrite it from scratch

The new keygen has the following behavior:
* generate just 16 notes at a time, each 1/16th in duration.
* pick notes via random sample, where the tonic has a weight of 2, and
  all other notes in the scale have a weight of 1.
* `--mode`, `--tonic`, and `DEST` are required arguments.
* installable like a normal local python package.